fix: searchParams 에러 해결 & svg 태그 Image(next) -> img test
