+++ src/app/(main)/search/filters/page.tsx
@@ -14,7 +14,7 @@ export default async function FiltersPage({
   const cat = (catParam ?? null) as CategoryKey | null;
 
   return (
-    <Suspense fallback={<div className="p-6">로딩 중…</div>}>
+    <Suspense fallback={null}>
       <SearchPage initialCat={cat} />
     </Suspense>
   );
